Adds utils for logging and toasts Refractors name async package to network
