+++ app/src/main/java/in/silive/directme/Utils/Constants.java
@@ -1,0 +1,9 @@
+package in.silive.directme.Utils;
+
+/**
+ * Created by shobhit on 28/2/17.
+ */
+
+public class Constants {
+    public static final String LOG = "debugging";
+}
