+++ app/src/main/java/in/silive/directme/Utils/LoggerUtils.java
@@ -1,0 +1,17 @@
+package in.silive.directme.Utils;
+
+import android.util.Log;
+
+import in.silive.directme.Activity.ParkedActivity;
+
+/**
+ * Created by shobhit on 28/2/17.
+ */
+
+public class LoggerUtils {
+
+    public static void logger(String message)
+    {
+        Log.d(Constants.LOG, message);
+    }
+}
