+++ app/src/main/java/in/silive/directme/application/DirectMe.java
@@ -1,6 +1,7 @@
 package in.silive.directme.application;
 
 import android.app.Application;
+import android.content.Context;
 
 /**
  * Created by shobhit on 27/2/17.
@@ -9,8 +10,9 @@ import android.app.Application;
 public class DirectMe extends Application {
 
     private static DirectMe singleton = null;
+    public Context mContext;
 
-    public static DirectMe getInstance(){
+    public static DirectMe getInstance() {
         return singleton;
     }
